Introduced Grids and separated tiles from the Board.
A new class template Grid is introduced to handle two-dimensional
arrays. The elements of the arrays can be accessed in a neat way using
Line and Segment templates. Using these, the code for computing a
move's score has been greatly simplified.  The placement of tiles in a
given game state has been separated from the Board and moved into a
new class, TilePlacement, which is a Grid of shared-pointers to
tiles. This allows to avoid copying the board's fields into each game
state.